django-api/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
PyJWT==2.8.0
dj-database-url==2.1.0
gunicorn==21.2.0
django-cors-headers==4.3.1
